\name{vagam-package}
\alias{vagam-package}
\docType{package}
\title{
Variational approximations for generalized additive models
}
\description{
\packageDescription{vagam}
}
\author{
\packageAuthor{vagam}
Maintainer: \packageMaintainer{vagam}
}
\references{
\itemize{
\item Hui, F. K. C., You, C., Shang, H. L., and Mueller, S. (2018). Semiparametric regression using variational approximations, \emph{Journal of the American Statistical Association}, \bold{forthcoming}.
} }
\keyword{package}

\examples{
## Please see examples in the help file for the vagam function.
}
